MongoDBHelper split in 2 classes: new MongoConfiguration handles connections
